% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mae_function.R
\name{mae}
\alias{mae}
\title{Mean Absolute Error}
\usage{
mae(error)
}
\description{
This function returns Mean Absolute Error
}
\examples{
mae()
}
\keyword{absolute}
\keyword{error}
\keyword{evaluation}
\keyword{mae}
\keyword{mean}
\keyword{model}
